# Revert optimization of constants and class vars