As I explain [here](https://forum.crystal-lang.org/t/specs-on-crystal-0-36-0-are-taking-a-long-time/2910/15?u=asterite), it seems these changes are the reason the 0.36.0 compiler is so slow.

Reverts #9801
Reverts #9995

I didn't use `git revert` because then an optimization to not use an "init" flag for string constants came in, and that's fine.

Fixes https://forum.crystal-lang.org/t/specs-on-crystal-0-36-0-are-taking-a-long-time/2910