@kostya 2021-01-29
I not think this was small optimization, some critical code can be 3.4 times faster:

little bench from real code:
```crystal
COMBS = UInt8[1, 1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0,
  0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,
  0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1,
  0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0].to_unsafe

INVERTED = UInt8[10, 9, 8, 11, 6, 5, 4, 11, 2, 1, 0, 7, 14, 13, 12, 15].to_unsafe

@[AlwaysInline]
private def combo_index(v : UInt8, other : UInt8) : UInt8
  (v << 4) | other
end

@[AlwaysInline]
def good?(v : UInt8, other : UInt8) : Bool
  COMBS[combo_index(v, INVERTED[other])] != 0
end

N = (ARGV[0]? || 100000000).to_i

t = Time.local
c = 0_u64
N.times do |i|
  d1 = i & 15
  d2 = rand(16)
  c += 1 if good?(d1.to_u8, d2.to_u8)
end
p c
p Time.local - t

```

crystal 0.35.1 -> 0.86s
crystal 0.36.0 -> 0.25s