@asterite 2021-01-29
I see. We can always reintroduce this optimization later while also making sure we don't make compile times incredibly slow. But right now the priority is to release 0.36.1 soon.