@j8r 2021-01-29
Couldn't the optimizarion only be done in release mode?