@asterite 2021-01-29
@j8r Well, macro runs are compiled in release mode, so...