@j8r 2021-01-29
Or maybe like for gcc/clang with `-O2`, or an other compile-time optimization flag :smile: 
Not needed if it is possible to have both a fast program and compile fast as you said.

However, not sure it will be always the case in the future...