@bcardiff 2021-01-29
At this point is reverting to get a reasonable experience. The motivation for the optimization still holds but the current efforts were not enough. Once they are worked out fully they can come back. It's not that bad :-)