@mattrberry 2021-01-29
I agree that the performance improvements here likely aren't worth the compile-time hit. I'd also just like to throw my perspective in. These optimizations had bought me 15%-20% better performance for CryBoy on 0.36.0 vs 0.35.1, which is pretty significant. Even with this rollback, it'd be nice to find some other way to bring these optimizations back at a later point