@asterite 2021-01-29
I still don't know why LLVM takes a lot more time to optimize this.

One thing I can try is to move the initialization of each constant/class_var to a separate function. Maybe like that it's easier for LLVM, I don't know. I have to try it.

But we can always try that later, before or after 1.0. Though if I have time before the next patch release, I'll do it!