@asterite 2021-01-29
I have a separate PR where the optimization remains and the compile times are not affected (I think) 😸